rename: commonTestState -> reviewState로 파일명 변경
